+++ lib/libgtkwave/src/gw-loader.c
@@ -7,7 +7,6 @@ typedef struct
     gboolean autocoalesce;
 
     gchar hierarchy_delimiter;
-    gchar alternate_hierarchy_delimiter;
 
     gboolean already_used;
 } GwLoaderPrivate;
@@ -20,7 +19,6 @@ enum
     PROP_PRESERVE_GLITCHES_REAL,
     PROP_AUTOCOALESCE,
     PROP_HIERARCHY_DELIMITER,
-    PROP_ALTERNATE_HIERARCHY_DELIMITER,
     N_PROPERTIES,
 };
 
@@ -48,10 +46,6 @@ static void gw_loader_set_property(GObject *object,
 
         case PROP_HIERARCHY_DELIMITER:
             gw_loader_set_hierarchy_delimiter(self, g_value_get_uchar(value));
-            break;
-
-        case PROP_ALTERNATE_HIERARCHY_DELIMITER:
-            gw_loader_set_alternate_hierarchy_delimiter(self, g_value_get_uchar(value));
             break;
 
         default:
@@ -82,10 +76,6 @@ static void gw_loader_get_property(GObject *object,
 
         case PROP_HIERARCHY_DELIMITER:
             g_value_set_uchar(value, gw_loader_get_hierarchy_delimiter(self));
-            break;
-
-        case PROP_ALTERNATE_HIERARCHY_DELIMITER:
-            g_value_set_uchar(value, gw_loader_get_alternate_hierarchy_delimiter(self));
             break;
 
         default:
@@ -130,16 +120,6 @@ static void gw_loader_class_init(GwLoaderClass *klass)
                            ' ',
                            127,
                            '.',
-                           G_PARAM_READWRITE | G_PARAM_EXPLICIT_NOTIFY | G_PARAM_STATIC_STRINGS);
-
-    // TODO: change type to unicode character
-    properties[PROP_ALTERNATE_HIERARCHY_DELIMITER] =
-        g_param_spec_uchar("alternate-hierarchy-delimiter",
-                           NULL,
-                           NULL,
-                           0,
-                           127,
-                           0,
                            G_PARAM_READWRITE | G_PARAM_EXPLICIT_NOTIFY | G_PARAM_STATIC_STRINGS);
 
     g_object_class_install_properties(object_class, N_PROPERTIES, properties);
@@ -273,26 +253,4 @@ gchar gw_loader_get_hierarchy_delimiter(GwLoader *self)
     GwLoaderPrivate *priv = gw_loader_get_instance_private(self);
 
     return priv->hierarchy_delimiter;
-}
-
-void gw_loader_set_alternate_hierarchy_delimiter(GwLoader *self, gchar delimiter)
-{
-    g_return_if_fail(GW_IS_LOADER(self));
-
-    GwLoaderPrivate *priv = gw_loader_get_instance_private(self);
-
-    if (priv->alternate_hierarchy_delimiter != delimiter) {
-        priv->alternate_hierarchy_delimiter = delimiter;
-
-        g_object_notify_by_pspec(G_OBJECT(self), properties[PROP_ALTERNATE_HIERARCHY_DELIMITER]);
-    }
-}
-
-gchar gw_loader_get_alternate_hierarchy_delimiter(GwLoader *self)
-{
-    g_return_val_if_fail(GW_IS_LOADER(self), FALSE);
-
-    GwLoaderPrivate *priv = gw_loader_get_instance_private(self);
-
-    return priv->alternate_hierarchy_delimiter;
 }