+++ src/file.h
@@ -10,7 +10,6 @@
 #ifndef WAVE_FILESEL_H
 #define WAVE_FILESEL_H
 
-void fileselbox_old(const char *title, char **filesel_path, GCallback ok_func, GCallback notok_func, const char *pattn, int is_writemode);
 void fileselbox(const char *title, char **filesel_path, GCallback ok_func, GCallback notok_func, const char *pattn, int is_writemode);
 
 #endif