+++ src/main.c
@@ -194,9 +194,8 @@ static void close_all_fst_files(void) /* so mingw does delete of reader tempfile
 {
     unsigned int i;
     for (i = 0; i < GLOBALS->num_notebook_pages; i++) {
-        if ((*GLOBALS->contexts)[i]->fst_fst_c_1) {
-            fstReaderClose((*GLOBALS->contexts)[i]->fst_fst_c_1);
-            (*GLOBALS->contexts)[i]->fst_fst_c_1 = NULL;
+        if ((*GLOBALS->contexts)[i]->dump_file) {
+            g_clear_object(&((*GLOBALS->contexts)[i])->dump_file);
         }
     }
 }
@@ -293,7 +292,7 @@ static void print_help(char *nam)
 
     printf(
         "Usage: %s [OPTION]... [DUMPFILE] [SAVEFILE] [RCFILE]\n\n"
-        "  -n, --nocli=DIRPATH        use file requester for dumpfile name\n"
+        "  -n, --nocli=DIRPATH        use file selection dialog for dumpfile or savefile\n"
         "  -f, --dump=FILE            specify dumpfile name\n" VCD_GETOPT
         "  -a, --save=FILE            specify savefile name\n"
         "  -r, --rcfile=FILE          specify override .rcfile name\n"
@@ -360,13 +359,12 @@ static char *wave_get_filename(char *dfile)
         }
 #endif
     }
-    fileselbox_old("GTKWave: Select a dumpfile...",
-                   &GLOBALS->ftext_main_main_c_1,
-                   G_CALLBACK(wave_get_filename_cleanup),
-                   G_CALLBACK(wave_get_filename_cleanup),
-                   NULL,
-                   0);
-    gtk_main();
+    fileselbox("GTKWave: Select a dumpfile...",
+                &GLOBALS->ftext_main_main_c_1,
+                G_CALLBACK(wave_get_filename_cleanup),
+                G_CALLBACK(wave_get_filename_cleanup),
+                NULL,
+                0);
 
     return (GLOBALS->ftext_main_main_c_1);
 }
@@ -803,7 +801,6 @@ int main_2(int opt_vcd, int argc, char *argv[])
         GLOBALS->wave_scrolling = old_g->wave_scrolling;
         GLOBALS->do_zoom_center = old_g->do_zoom_center;
         GLOBALS->zoom_pow10_snap = old_g->zoom_pow10_snap;
-        GLOBALS->alt_hier_delimeter = old_g->alt_hier_delimeter;
         GLOBALS->cursor_snap = old_g->cursor_snap;
         GLOBALS->hier_delimeter = old_g->hier_delimeter;
         GLOBALS->hier_was_explicitly_set = old_g->hier_was_explicitly_set;