+++ src/tcl_np.c
@@ -66,10 +66,8 @@ extern int NpLoadLibrary(HMODULE *tclHandle, char *dllName, int dllNameSize, cha
 {
     char *envdll, libname[MAX_PATH];
     HMODULE handle = (HMODULE)NULL;
-
     char path[MAX_PATH], *p;
-    /* #include <windows.h> */
-    /* #include <iostream> */
+
     if (!GetModuleFileName(NULL, path, MAX_PATH)) {
         printf("GetModuleFileName() failed\n");
     } else {
@@ -129,7 +127,7 @@ extern int NpLoadLibrary(HMODULE *tclHandle, char *dllName, int dllNameSize, cha
         /*
          * Try based on ActiveTcl registry entry
          */
-        char path[MAX_PATH], vers[MAX_PATH];
+        char vers[MAX_PATH];
         DWORD result, size = MAX_PATH;
         HKEY regKey;
 #define TCL_REG_DIR_KEY "Software\\ActiveState\\ActiveTcl"
@@ -202,7 +200,7 @@ extern int NpLoadLibrary(HMODULE *tclHandle, char *dllName, int dllNameSize, cha
                                     (char *)&msgPtr,
                                     0,
                                     NULL);
-            NpLog3("GetModuleFileNameA ERROR: %d (%s)\n",
+            NpLog3("GetModuleFileNameA ERROR: %lu (%s)\n",
                    code,
                    ((length == 0) ? "unknown error" : msgPtr));
             if (length > 0) {